refactor: extract chat post-prompt pipeline
